requirements.txt
pipecat-ai[daily,cartesia,openai,silero,deepgram]
fastapi
uvicorn
aiohttp
pytz
python-dotenv
twilio
requests
loguru